# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample tiket_test/server Petstore tiket_test/server.",
        "title": "Tech Test Api Docs",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/flight/search": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Adding search queue",
                "parameters": [
                    {
                        "description": "register search",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.StdResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/delivery.SearchResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "delivery.SearchRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2025-07-10"
                },
                "from": {
                    "type": "string"
                },
                "passengers": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "delivery.SearchResponse": {
            "type": "object",
            "properties": {
                "search_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "helper.StdResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Just insert after you hit login",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}